+++ src/frontend/Sider/Behandling/Inngangsvilkår/Målgruppe/MålgruppeVilkår.tsx
@@ -4,7 +4,7 @@ import styled from 'styled-components';
 
 import { Alert, BodyLong, Heading } from '@navikt/ds-react';
 
-import { EndreMålgruppeForm } from './EndreMålgruppeRad';
+import { EndreMålgruppeForm } from './EndreMålgruppeGenerell';
 import { målgruppeTilMedlemskapHjelpetekst } from './hjelpetekstVurdereMålgruppe';
 import { målgrupperHvorMedlemskapMåVurderes, skalVurdereDekkesAvAnnetRegelverk } from './utils';
 import { JaNeiVurdering } from '../../Vilkårvurdering/JaNeiVurdering';
